server/app/requirements.txt
boto3
daphne
dj-database-url
django
django-cors-headers
django-storages
django-sass-processor
djangorestframework
channels
channels_redis
asgi-redis
factory_boy
libsass
pillow
psycopg2
pytest_django
redis
uwsgi
